Test zero-downtime deployment

+++ tests/handles_example.sh
@@ -11,10 +11,35 @@ test_container_engine() {
 
   docker compose down
 
-  "${WHEELSTICKS}" --container-engine "${container_engine}" deploy --wait \
-    --wait-timeout 30
+  "${WHEELSTICKS}" --container-engine "${container_engine}" \
+    deploy --wait --wait-timeout 30
 
-  curl --fail --max-time 0.2 --show-error http://localhost:8080
+  while true; do
+    curl --fail --max-time 0.2 --silent http://localhost:8080 || echo "Error $?"
+    sleep 0.01s
+  done >test.log &
+
+  sleep 2s
+
+  HI_VERSION=B "${WHEELSTICKS}" --container-engine "${container_engine}" \
+    deploy --wait --wait-timeout 30
+
+  sleep 2s
+
+  kill %%
+
+  for hi_version in 'A' 'B'; do
+    if ! grep --quiet "Hi from ${hi_version}" test.log; then
+      echo "No successful ping for hi version: ${hi_version}" >&2
+      exit 1
+    fi
+  done
+
+  local -r ping_errors="$(grep Error test.log)"
+  if [[ -n "${ping_errors}" ]]; then
+    printf 'Failed pings:\n%s\n' "${ping_errors}" >&2
+    exit 1
+  fi
 
   docker compose down
 }
